apps/server/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==42.0.5
cssselect==1.2.0
decorator==5.1.1
executing==2.0.1
fastapi==0.110.3
filelock==3.13.4
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
ipython==8.24.0
itemadapter==0.8.0
itemloaders==1.2.0
jedi==0.19.1
jmespath==1.0.1
lxml==5.2.1
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mysql-connector-python==8.4.0
packaging==24.0
parsel==1.9.1
parso==0.8.4
pexpect==4.9.0
prompt-toolkit==3.0.43
Protego==0.3.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyDispatcher==2.0.7
pyee==11.1.0
Pygments==2.17.2
pyOpenSSL==24.1.0
python-dotenv==1.0.1
PyYAML==6.0.1
queuelib==1.6.2
requests==2.31.0
requests-file==2.0.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
stack-data==0.6.3
starlette==0.37.2
tldextract==5.1.2
traitlets==5.14.3
Twisted==24.3.0
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
w3lib==2.1.2
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
zope.interface==6.3
